fix: drop the redundant i64 cast in the timezone test
The offset is already milliseconds as i64. Clippy rejects i64::from on
that value, which stopped the Linux, macOS, and Windows checks.

+++ native/opennow-core/src/cloudmatch.rs
@@ -4385,7 +4385,7 @@ mod tests {
             let offset = body["sessionRequestData"]["clientTimezoneOffset"]
                 .as_i64()
                 .unwrap();
-            assert!(offset == i64::from(offset_before) || offset == i64::from(offset_after));
+            assert!(offset == offset_before || offset == offset_after);
         }
     }
